refactor(testing): split the key-generation CLI out of the key manager
The key module mixed two unrelated concerns: the runtime signing
manager that every fixture and unit test imports, and a standalone
generate/download command-line tool that only runs from CI or a
developer shell. The tool section also carried mid-function imports
of sys and time, violating the top-of-file import rule.

Changes:

- the new keys_cli module holds the download URLs, the CLI-only
  defaults (validator count, generation max slot), the parallel
  key-generation workers, the archive download, and the entry point;
  its imports all live at the top of the file
- the key module keeps the signing manager, the dummy signature, the
  scheme map, and the key-directory resolver; its module docstring
  points at the companion CLI
- the missing-keys error hints and the CI workflows invoke the new
  module path; the fill CLI imports the download from it
